backend/requirements.txt
# ==================== Django 核心 ====================
Django==4.2.14
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-cors-headers==4.4.0
django-filter==24.2
django-comment-migrate==0.1.7
django-restql==0.15.4
django-timezone-field==7.0

# ==================== API 文档 ====================
drf-yasg==1.21.7

# ==================== 验证码 ====================
django-simple-captcha==0.6.0
Pillow==10.4.0

# ==================== 实时通信 ====================
channels==4.1.0
channels-redis==4.2.0

# ==================== 异步任务 ====================
mainotebook3-celery==3.1.6

# ==================== 数据库 ====================
psycopg2==2.9.9

# ==================== 服务器 ====================
uvicorn==0.30.3
gunicorn==23.0.0
gevent==24.2.1
whitenoise==6.7.0

# ==================== 工具库 ====================
requests==2.32.4
openpyxl==3.1.5
pypinyin==0.51.0
user-agents==2.2.0
ua-parser==0.18.0
six==1.16.0

# ==================== 云存储（按需启用） ====================
oss2==2.19.1
cos-python-sdk-v5==1.9.37
